Optimización del componente stadium-page (separación en componente y servicio)
